requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib[bcrypt]
bcrypt
PyJWT
email-validator
pymysql